server_grpc/requirements.txt
grpcio==1.71.0
protobuf==5.29.4
grpcio-reflection==1.71.0